requirements.txt
PyQt6>=6.4.0
SQLAlchemy>=2.0.0
python-magic>=0.4.27
ctransformers>=0.2.24
PyMuPDF>=1.22.5  # For PDF preview support
python-docx>=0.8.11  # For Word document preview
pandas>=2.0.0  # For Excel preview
openpyxl>=3.1.0  # Excel engine for pandas
Pillow>=9.5.0  # For image preview
mutagen>=1.45.1  # For audio file metadata
opencv-python>=4.7.0  # For video preview
